Refactor my way to an SPA
Mainly, rescue the SPA logic from overlay-pagination and establish a (god-like) controller that controls the main interactions and history-handling.

Also, fine-tune some colors and animations, completely adopt search into own component and treat it with a route handler and all.

Controller and the components need much more love and care. But I should finalize this single-commit beast for now.